feat: Add case existence check before inserting in Pennsylvania scraper

+++ src/scrapers/pennsylvania.py
@@ -127,9 +127,13 @@ class PennsylvaniaScraper(ScraperBase):
                             case_details = self.parse_case_details(content)
                             console.log(case_details)
                             case_details = cases_model.Case.model_validate(case_details)
-                            cases.insert_case(case_details)
-                            leads.insert_lead_from_case(case_details)
-                            console.log(f"inseted case {case_details.case_id}")
+                            case = cases.get_single_case(case_details.case_id)
+                            if case is None:
+                                cases.insert_case(case_details)
+                                leads.insert_lead_from_case(case_details)
+                                console.log(f"inseted case {case_details.case_id}")
+                            else: 
+                                console.log(f"Case {case_details.case_id} already exists.")
 
                 except Exception as e:
                     console.print(
@@ -137,7 +141,7 @@ class PennsylvaniaScraper(ScraperBase):
                     )
                 finally:
                     try:
-                        if os.path.exists(case_file_path):
+                        if case_file_path and  os.path.exists(case_file_path):
                             os.remove(case_file_path)
                             console.log(f"File {case_file_path} removed successfully")
                     except Exception as e:
